// client/src/constants/errors.enum.ts
export enum FormErrors {
  INVALID_EMAIL = "Please enter a valid email address",
  EMAIL_REQUIRED = "Email is required",
  PASSWORD_FORMAT = "Password must contain at least one letter, one number, and one special character.",
  PASSWORD_LENGTH = "Password must be at least 8 characters long",
  PASSWORD_REQUIRED = "Password is required",
  NAME_MIN_LENGTH = "Name must be at least 3 characters long",
  NAME_MAX_LENGTH = "Name cannot exceed 96 characters",
  NAME_REQUIRED = "Name is required",
}